+++ components/Map/MapGL.tsx
@@ -37,6 +37,8 @@ import {
   useUnlikeMutation,
 } from '@graphql/relations.graphql';
 import { query } from 'express';
+import UpdateUrl from './UpdateUrl';
+import { motion, AnimatePresence } from 'framer-motion';
 
 type PlaceProps = {
   id: number;
@@ -63,31 +65,7 @@ const GetBounds = (bounds: {
   };
 };
 
-type UpdateUrlProps = {
-  longitude: number;
-  latitude: number;
-  zoom: number;
-  router: NextRouter;
-};
-
-// save latitude, longitude, zoom in url props
-const UpdateUrl = ({
-  router,
-  longitude,
-  latitude,
-  zoom,
-}: UpdateUrlProps): void => {
-  router.replace({
-    pathname: 'map',
-    query: {
-      lat: latitude.toFixed(6),
-      lng: longitude.toFixed(6),
-      zoom: zoom.toFixed(2),
-    },
-  });
-};
-
-const MapGL: FC<{
+type MapGLProps = {
   searchCoordinates: { lat: number; lng: number };
   oldPoints: Array<PlaceProps>;
   setBounds: React.Dispatch<
@@ -98,13 +76,14 @@ const MapGL: FC<{
       minLongitude: number;
     }>
   >;
-}> = ({ searchCoordinates, setBounds, oldPoints }) => {
+};
+
+const MapGL: FC<MapGLProps> = ({ searchCoordinates, setBounds, oldPoints }) => {
   const router = useRouter();
   const [timeLimitation, setTimeLimitation] = useState<[number, number]>([
     0,
     new Date().getTime(),
   ]);
-
   const paths = usePathsQuery();
 
   const [viewport, setViewport] = useState({
@@ -282,45 +261,47 @@ const MapGL: FC<{
           showCompass={false}
         />
 
-        {clusters.map((cluster) => {
-          const [longitude, latitude] = cluster.geometry.coordinates;
-          const { cluster: isCluster, point_count: pointCount } =
-            cluster.properties;
+        <AnimatePresence>
+          {clusters.map((cluster) => {
+            const [longitude, latitude] = cluster.geometry.coordinates;
+            const { cluster: isCluster, point_count: pointCount } =
+              cluster.properties;
 
-          if (isCluster) {
-            return (
-              <Marker
-                key={`cluster-${cluster.id}`}
-                latitude={latitude}
-                longitude={longitude}
-              >
-                <div
-                  className="text-white bg-red-500 h-20 w-20 rounded-full py-[2em] text-center "
-                  onClick={() => {
-                    const expansionZoom = Math.min(
-                      supercluster.getClusterExpansionZoom(cluster.id),
-                      20
-                    );
-
-                    setViewport({
-                      ...viewport,
-                      latitude,
-                      longitude,
-                      zoom: expansionZoom,
-                    });
-                  }}
+            if (isCluster) {
+              return (
+                <Marker
+                  key={`cluster-${cluster.id}`}
+                  latitude={latitude}
+                  longitude={longitude}
                 >
-                  {pointCount}
-                </div>
-              </Marker>
+                  <div
+                    className="text-white bg-red-500 h-20 w-20 rounded-full py-[2em] text-center "
+                    onClick={() => {
+                      const expansionZoom = Math.min(
+                        supercluster.getClusterExpansionZoom(cluster.id),
+                        20
+                      );
+
+                      setViewport({
+                        ...viewport,
+                        latitude,
+                        longitude,
+                        zoom: expansionZoom,
+                      });
+                    }}
+                  >
+                    {pointCount}
+                  </div>
+                </Marker>
+              );
+            }
+            const place = filteredPlaces.find(
+              (place: PlaceProps) =>
+                place.latitude === latitude && place.longitude === longitude
             );
-          }
-          const place = filteredPlaces.find(
-            (place: PlaceProps) =>
-              place.latitude === latitude && place.longitude === longitude
-          );
-          return place ? <MapPlace key={cluster.id} place={place} /> : null;
-        })}
+            return place ? <MapPlace key={cluster.id} place={place} /> : null;
+          })}
+        </AnimatePresence>
       </ReactMapGL>
       <div className="absolute left-[10vw] top-20 z-50 w-[80vw]">
         <TimeLine
@@ -338,11 +319,16 @@ const MapPlace = ({ place }: { place: any }) => {
   return (
     <>
       <Marker latitude={place.latitude} longitude={place.longitude}>
-        <a
-          className=""
-          onClick={() => {
-            setIsOpen(true);
+        {/* marker appear and disappear animation */}
+        <motion.div
+          initial={{ scale: 0.4, opacity: 0 }}
+          animate={{ scale: 1, opacity: 1 }}
+          exit={{ scale: 0.4, opacity: 0 }}
+          transition={{
+            ease: 'easeOut',
+            duration: 0.2,
           }}
+          onClick={() => setIsOpen(true)}
         >
           <Image
             src={place.posts[0].url[0]}
@@ -351,20 +337,24 @@ const MapPlace = ({ place }: { place: any }) => {
             className="rounded-full w-12 h-12 object-cover"
             alt="Picture on map"
           />
-        </a>
+        </motion.div>
       </Marker>
-      <MapModal setIsOpen={setIsOpen} isOpen={isOpen} place={place} />
+      <PlaceModal setIsOpen={setIsOpen} isOpen={isOpen} place={place} />
     </>
   );
 };
 
-type MapModalProps = {
+type PlaceModalProps = {
   isOpen: boolean;
   setIsOpen: React.Dispatch<React.SetStateAction<boolean>>;
   place: any;
 };
 
-const MapModal: React.FC<MapModalProps> = ({ isOpen, setIsOpen, place }) => {
+const PlaceModal: React.FC<PlaceModalProps> = ({
+  isOpen,
+  setIsOpen,
+  place,
+}) => {
   const photos = place.posts.map((photo: { url: string[] }) => ({
     src: photo.url[0],
     width: 4,
@@ -381,17 +371,29 @@ const MapModal: React.FC<MapModalProps> = ({ isOpen, setIsOpen, place }) => {
 
   return (
     <>
-      <Modal
-        open={isOpen}
-        onClose={() => {
-          setDetail(false);
-          setIsOpen(false);
-        }}
-      >
-        <section className="absolute justify-between z-50 top-[50%] left-[50%] bg-white max-h-[600px] max-w-[800px] w-full h-full -translate-y-1/2 -translate-x-1/2 rounded-xl">
-          <Gallery photos={photos} onClick={openLightbox} />
-        </section>
-      </Modal>
+      <AnimatePresence>
+        <motion.section
+          initial={{ scale: 0.4, opacity: 0 }}
+          animate={{ scale: 1, opacity: 1 }}
+          exit={{ scale: 0.4, opacity: 0 }}
+          transition={{
+            ease: 'easeOut',
+            duration: 0.2,
+          }}
+        >
+          <Modal
+            open={isOpen}
+            onClose={() => {
+              setDetail(false);
+              setIsOpen(false);
+            }}
+          >
+            <section className="absolute justify-between z-50 top-[50%] left-[50%] bg-white max-h-[600px] max-w-[800px] w-full h-full -translate-y-1/2 -translate-x-1/2 rounded-xl">
+              <Gallery photos={photos} onClick={openLightbox} />
+            </section>
+          </Modal>
+        </motion.section>
+      </AnimatePresence>
       <DetailModal
         open={detail}
         onClose={() => {
@@ -401,7 +403,7 @@ const MapModal: React.FC<MapModalProps> = ({ isOpen, setIsOpen, place }) => {
         place={place}
         currentImage={currentImage}
         setCurrentImage={setCurrentImage}
-      />
+      />{' '}
     </>
   );
 };
