+++ package.json
@@ -54,7 +54,7 @@
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
     "formik": "^2.2.9",
-    "framer-motion": "^4.1.17",
+    "framer-motion": "^5.0.0",
     "graphql": "^15.6.0",
     "graphql-let": "^0.18.5",
     "graphql-upload": "^12.0.0",
